# core: optimized the time it takes to decrypt all filemetadata