In this PR, I created a hashmap that stores AESKeys whilst file metadata is being decrypted. This means these keys can be reused to decrypt subsequent children.


An instance of this improvement with `lockbook list`.
Before:
```
./lockbook list  0.10s user 0.01s system 98% cpu 0.110 total
```

After:
```
./lockbook list  0.01s user 0.01s system 98% cpu 0.018 total
```